feat: create EnderecoRepository for Endereco entity

+++ server/src/main/java/br/com/moria/repositories/MembroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.moria.models.Membro;
 
-public interface MembroRepository extends JpaRepository<Membro, Long> {
+public interface MembroRepository extends JpaRepository<Membro, Integer> {
 
 }
